upath._flavour: factor out FSSpecFlavour
